Fixed legend output (SVG + EPS) Improved graph/legend layout centering

+++ src/CORE/legend.cpp
@@ -8,7 +8,7 @@ Legend::Legend(const ColorMap *map)
 	_size = 0;
 }
 
-void Legend::setSize(size_t size)
+void Legend::setSize(int size)
 {
 	_size = size;
 	computeDimensions();
@@ -21,7 +21,7 @@ void Legend::updateMap()
 
 void Legend::computeDimensions()
 {
-	size_t height = LEGEND_MARGIN, tl = 0;
+	int height = LEGEND_MARGIN, tl = 0;
 	float width;
 
 	if (!_size) {
@@ -30,12 +30,12 @@ void Legend::computeDimensions()
 	}
 
 	for (ColorMap::const_iterator it = _map->begin(); it != _map->end(); it++) {
-		tl = MAX(tl, (*it).first.length());
+		tl = MAX(tl, (int)((*it).first.length()));
 		height += _size * LEGEND_RECT_RATIO;
 	}
 
 	width = LEGEND_MARGIN + (_size * LEGEND_RECT_RATIO) + (_size / 3)
 	  + (_size * AVG_CHAR_WIDTH * tl);
 
-	_dimensions = Coordinates((size_t)width, height);
+	_dimensions = Coordinates((int)width, height);
 }
